requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.27.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
python-decouple==3.8
google-generativeai==0.3.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
elasticsearch==8.11.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0